загрузка задач из файла вынесена в отдельны класс TaskFormatter
